Revert back to a 16-bit DDR3 bus, initial DDR3 layout.
The caBGA-554 package was too large and fat to reasonably be able
to fit within the beaglebone footprint and still leave enough room
for DDR3 and the expansion headers. So, it seems more sensible to
switch back to a 16-bit wide bus.

Now that we can fit everything, take a first stab at getting the DDR3
data busses routed and length matched.