Replace PMIC with MXL7704, switch to SPI master programming.
The Quorvo PMIC in the original design appears to have been deprecated
on DigiKey. Out of concern for availability, let's switch to the MXL7704
instead.

In an effort to support to an FPGA-based bootloader solution, we have
switched to using SPI master mode for the ECP5 configuration data. In
the long run this might allow a considerable downsizing of the STM32.