Add dedicated LDO for VCCAUX, and a bunch of cleanup
The MXL7704 is almost capable of booting the entire ECP5 on its own
without any help from the supervisor. So we would like to test it out
on the prototype boards. To accompish this, we need to use a dedicated
LDO for VCCAUX.

With that, I am rapidly running out of major work to do and from here
on out it's just going to be review and minor tweaks.