Tweak BOM to use ECP5-5G for all variants.
The price difference between the LFE5UM-45F and the LFE5UM5G-45F is
only a few dollars, and enables the user to play with USB Superspeed
and PCIe 2.0, so this just seems like an obvious upgrade. We also
increase the choice of RAM for the base model to 1GB.

This also leads to a couple of tweaks in the LDO selection to switch
the core and SERDES voltages from 1.1V to 1.2V for the 5G models.