Add CERN OHL required documentation, annotate schematic.
The CERN OHL requires the inclusion of a CHANGES.md and a PRODUCT.txt
file to denote the change history of the project and the contact person
for production information. While we're at it, we also try to imrpove
the state of the schematic with explicit copyright and authorship notices
plus some of the bodges that were added during prototype bringup.